requirements.txt
flask
python-dotenv
azure-cosmos
